ai-model-api/pyproject.toml
[project]
name = "aurora-django-backend"
version = "0.1.0"
description = "Aurora Organics Django Backend"
dependencies = [
    
    # Flask Core
    "Flask>=3.0.0",
    "Flask-CORS>=4.0.0",
        
    # Machine Learning & AI
    "tensorflow>=2.13.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",

        # Image Processing
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
        
    # Data Processing
    "pandas>=2.0.0",
    
    # Utilities
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    
    "gunicorn>=21.2.0",
]
requires-python = ">=3.10"

[build-system]
requires    = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# <-- tell Hatch which package directory to include
packages = ["backend"]
